Revert "Yet cleaner"
This reverts commit a54bca2f205645a36092f32b5f6149b9be0b01bb.

+++ ciri/consensus/snapshot/local_snapshots/local_snapshots_manager.c
@@ -5,16 +5,16 @@
  * Refer to the LICENSE file for licensing information
  */
 
-#include <inttypes.h>
-
-#include "ciri/consensus/snapshot/local_snapshots/conf.h"
 #include "ciri/consensus/snapshot/local_snapshots/local_snapshots_manager.h"
+#include <inttypes.h>
+#include "ciri/consensus/snapshot/local_snapshots/conf.h"
 #include "ciri/consensus/snapshot/snapshots_service.h"
 #include "utils/logger_helper.h"
 #include "utils/macros.h"
 #include "utils/time.h"
 
 #define LOCAL_SNAPSHOTS_RESCAN_INTERVAL_MS 10000
+
 #define LOCAL_SNAPSHOTS_MANAGER_LOGGER_ID "local_snapshots_manager"
 
 static logger_id_t logger_id;
@@ -94,8 +94,7 @@ cleanup:
 
 bool iota_local_snapshots_manager_should_take_snapshot(local_snapshots_manager_t const *const lsm,
                                                        tangle_t const *const tangle) {
-  uint64_t new_transactions_count;
-
+  size_t new_transactions_count;
   if (iota_tangle_transaction_count(tangle, &new_transactions_count) != RC_OK) {
     log_critical(logger_id, "Failed in querying db size\n");
     return false;
